requirements.txt
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.1.0
autopep8==1.5.4
colorama==0.4.3
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
